USER - 내 쿠폰, 회원탈퇴, 주문결제, 주문, 리뷰

+++ src/main/java/kr/co/jhta/app/delideli/user/board/service/BoardService.java
@@ -31,7 +31,7 @@ public interface BoardService {
     List<Board> getAllKeywordEvent(int startNo, int endNo, String keyword);
 
     //내문의 목록
-    List<Board> getMyAskList(Long userKey);
+    List<Board> getMyAskList(int userKey);
     //내문의 글작성
     void myAskWrite(Board board);
     //내문의 상세보기
